5. building node.js server with express
setting up server dependencies in server.js. server side code is located in src/server.js. so there we import express, starting it on PORT or 3001 by default, and it response for any request with 'Hello World!'.

also nodemon was installed for compiling server side js and adding script to package.json to start it