2.2.1. adding pages and server side routing
On the server side express transfers request URL to the match function from react-router and it returns renderProps which we can use to render requested user content or we get that there are no correspondings and return the 404 page.

Learning App.jsx to render children components is crucial.

react-router-bootstrap is a cool stuff.