2.3.1. creating CountersPage before implementing Flux/Redux
StateCounter.jsx - parent component for Counter, it stores Counters current value in its own storage of state and business-logic for updating this value; implementing both incrementing and decrementing.

Counter.jsx - view component.

Currently if we click on the button and switch to another page and then come back the value of the Counter is not being saved (aka that's what Flux is for).