2.3.3. Redux initialisation
combineReducers combines all the reducers into one, so we can split global state into as many as needed small reducers which are independent between each other and are responsible only for their own business-logic.

For the whole stores initialisation we pass an empty object as an initial state.

Thunk is a middleware (assembly machine?) which is responsible for functions execution in dispatch.

createStore execution results with a store of a global application state which can process actions and can be changed due to its results.