3.3. adding global state to client side
After the server finished getting contect for the client we execute getState() on store to get current global state. Then we force it to the HTML template and return it to the browser. Client side JS gets global state right out the global window object and transfers it to the configureStore as an initialState