3.4. removing timeAction server 'dummy'
Fetch function is an extended isomorphic-fetch from redux-oauth. It must be called through dispatch so it has an access to the global store where it can get authorized token and send it with the request. If fetch is used for arbitrary HTTP-request but not an API request, then authorized token is not used.